steps miles put into routes form correctly

+++ app/src/main/java/com/example/walkwalkrevolution/RecyclerViewAdapter.java
@@ -178,11 +178,11 @@ public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerViewAdapte
         return "Date: " + date;
     }
 
-    private String formatSteps(int steps){
+    private String formatSteps(long steps){
         return "Steps: " + steps;
     }
 
-    private String formatMiles(float miles){
+    private String formatMiles(double miles){
         return "Miles: " + miles;
     }
